refactor: split focused frontend tests
